# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToDo API",
        "description": "API Documentation for ToDo Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/Name"
                    },
                    {
                        "$ref": "#/components/parameters/Description"
                    },
                    {
                        "$ref": "#/components/parameters/CategoryId"
                    },
                    {
                        "$ref": "#/components/parameters/PriorityId"
                    },
                    {
                        "$ref": "#/components/parameters/StatusId"
                    },
                    {
                        "$ref": "#/components/parameters/DueDateFrom"
                    },
                    {
                        "$ref": "#/components/parameters/DueDateTo"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "tasks": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Task"
                                                            }
                                                        },
                                                        "total_tasks": {
                                                            "type": "integer"
                                                        },
                                                        "total_pages": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task by ID",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TaskApiResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                        "properties": {
                            "data": {
                                "properties": {
                                    "": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "TaskRequest": {
                "required": [
                    "category_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "description": "ID of the task category",
                        "type": "integer",
                        "example": 1
                    },
                    "status_id": {
                        "description": "ID of the task status",
                        "type": "integer",
                        "example": 1
                    },
                    "priority_id": {
                        "description": "ID of the task priority",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the task",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Task Name"
                    },
                    "description": {
                        "description": "Description of the task",
                        "type": "string",
                        "example": "Task Description"
                    },
                    "text": {
                        "description": "Detailed text of the task",
                        "type": "string",
                        "example": "Detailed Task Text"
                    },
                    "due_date": {
                        "description": "Due date of the task",
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-31"
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "status_id": {
                        "type": "integer"
                    },
                    "priority_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Number of tasks per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                }
            },
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number for pagination",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "Name": {
                "name": "name",
                "in": "query",
                "description": "Name of the task",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "Description": {
                "name": "description",
                "in": "query",
                "description": "Description of the task",
                "required": false,
                "schema": {
                    "type": "string",
                    "maxLength": 255
                }
            },
            "CategoryId": {
                "name": "category_id",
                "in": "query",
                "description": "ID of the task category",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "PriorityId": {
                "name": "priority_id",
                "in": "query",
                "description": "ID of the task priority",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "StatusId": {
                "name": "status_id",
                "in": "query",
                "description": "ID of the task status",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            },
            "DueDateFrom": {
                "name": "due_date_from",
                "in": "query",
                "description": "Filter tasks with due date from",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            },
            "DueDateTo": {
                "name": "due_date_to",
                "in": "query",
                "description": "Filter tasks with due date to",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to authorize requests",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}